backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
pydantic
psycopg2-binary